+++ packages/vchart-theme-demo-component/CHANGELOG.md
@@ -1,6 +1,13 @@
 # Change Log - @visactor/vchart-theme-demo-component
 
-This log was last generated on Wed, 19 Jun 2024 11:37:42 GMT and should not be manually modified.
+This log was last generated on Mon, 05 Aug 2024 08:45:51 GMT and should not be manually modified.
+
+## 1.11.5
+Mon, 05 Aug 2024 08:45:51 GMT
+
+### Updates
+
+- feat: add business funnel demo
 
 ## 1.11.3
 Wed, 19 Jun 2024 11:37:42 GMT